chore(flags): deduplicate PushGatewayTaskTestMixin and add missing metric assertions
Extract PushGatewayTaskTestMixin to shared test utility and add
pushgateway metric assertions for team_metadata and flag_definitions
verification tasks.
