+++ posthog/tasks/test/test_flag_definitions_cache_tasks.py
@@ -2,29 +2,11 @@ from unittest.mock import MagicMock, patch
 
 from django.test import TestCase
 
-from prometheus_client import CollectorRegistry
-
 from posthog.tasks.feature_flags import (
     cleanup_stale_flag_definitions_expiry_tracking_task,
     refresh_expiring_flag_definitions_cache_entries,
 )
-
-
-class PushGatewayTaskTestMixin:
-    """Sets up mocked PushGateway context so PushGatewayTask-based tasks can run in tests."""
-
-    def setUp(self) -> None:
-        super().setUp()  # type: ignore[misc]
-        self.registry = CollectorRegistry()
-        self.mock_context = MagicMock()
-        self.mock_context.__enter__ = MagicMock(return_value=self.registry)
-        self.mock_context.__exit__ = MagicMock(return_value=False)
-        self.patcher = patch("posthog.tasks.utils.pushed_metrics_registry", return_value=self.mock_context)
-        self.patcher.start()
-
-    def tearDown(self) -> None:
-        self.patcher.stop()
-        super().tearDown()  # type: ignore[misc]
+from posthog.tasks.test.utils import PushGatewayTaskTestMixin
 
 
 class TestRefreshExpiringFlagDefinitionsCacheEntries(PushGatewayTaskTestMixin, TestCase):
