+++ posthog/tasks/test/test_hypercache_verification.py
@@ -11,30 +11,12 @@ from unittest.mock import MagicMock, patch
 
 from django.test import TestCase, override_settings
 
-from prometheus_client import CollectorRegistry
-
 from posthog.tasks.hypercache_verification import (
     verify_and_fix_flag_definitions_cache_task,
     verify_and_fix_flags_cache_task,
     verify_and_fix_team_metadata_cache_task,
 )
-
-
-class PushGatewayTaskTestMixin:
-    """Sets up mocked PushGateway context so PushGatewayTask-based tasks can run in tests."""
-
-    def setUp(self) -> None:
-        super().setUp()  # type: ignore[misc]
-        self.registry = CollectorRegistry()
-        self.mock_context = MagicMock()
-        self.mock_context.__enter__ = MagicMock(return_value=self.registry)
-        self.mock_context.__exit__ = MagicMock(return_value=False)
-        self.patcher = patch("posthog.tasks.utils.pushed_metrics_registry", return_value=self.mock_context)
-        self.patcher.start()
-
-    def tearDown(self) -> None:
-        self.patcher.stop()
-        super().tearDown()  # type: ignore[misc]
+from posthog.tasks.test.utils import PushGatewayTaskTestMixin
 
 
 @override_settings(FLAGS_REDIS_URL="redis://test")
@@ -92,7 +74,7 @@ class TestVerifyAndFixFlagsCacheTaskDisabled(TestCase):
 
 
 @override_settings(FLAGS_REDIS_URL="redis://test")
-class TestVerifyAndFixTeamMetadataCacheTask(TestCase):
+class TestVerifyAndFixTeamMetadataCacheTask(PushGatewayTaskTestMixin, TestCase):
     @patch("posthog.tasks.hypercache_verification._run_verification_for_cache")
     def test_verifies_team_metadata_cache(self, mock_run_verification: MagicMock) -> None:
         mock_run_verification.return_value = MagicMock()
@@ -124,6 +106,19 @@ class TestVerifyAndFixTeamMetadataCacheTask(TestCase):
 
         mock_run_verification.assert_called_once()
 
+    @patch("posthog.tasks.hypercache_verification._run_verification_for_cache")
+    def test_pushgateway_metrics_recorded_on_success(self, mock_run_verification: MagicMock) -> None:
+        mock_run_verification.return_value = MagicMock()
+
+        verify_and_fix_team_metadata_cache_task()
+
+        success = self.registry.get_sample_value("posthog_celery_verify_and_fix_team_metadata_cache_task_success")
+        duration = self.registry.get_sample_value(
+            "posthog_celery_verify_and_fix_team_metadata_cache_task_duration_seconds"
+        )
+        assert success == 1
+        assert duration is not None and duration >= 0
+
 
 @override_settings(FLAGS_REDIS_URL=None)
 class TestVerifyAndFixTeamMetadataCacheTaskDisabled(TestCase):
@@ -134,7 +129,7 @@ class TestVerifyAndFixTeamMetadataCacheTaskDisabled(TestCase):
         mock_run_verification.assert_not_called()
 
 
-class TestVerifyAndFixFlagDefinitionsCacheTask(TestCase):
+class TestVerifyAndFixFlagDefinitionsCacheTask(PushGatewayTaskTestMixin, TestCase):
     """Tests for the flag definitions cache verification task."""
 
     @patch("posthog.tasks.hypercache_verification._run_verification_for_cache")
@@ -158,6 +153,19 @@ class TestVerifyAndFixFlagDefinitionsCacheTask(TestCase):
 
         mock_capture.assert_called_once_with(error)
         assert context.exception is error
+
+    @patch("posthog.tasks.hypercache_verification._run_verification_for_cache")
+    def test_pushgateway_metrics_recorded_on_success(self, mock_run_verification: MagicMock) -> None:
+        mock_run_verification.return_value = MagicMock()
+
+        verify_and_fix_flag_definitions_cache_task()
+
+        success = self.registry.get_sample_value("posthog_celery_verify_and_fix_flag_definitions_cache_task_success")
+        duration = self.registry.get_sample_value(
+            "posthog_celery_verify_and_fix_flag_definitions_cache_task_duration_seconds"
+        )
+        assert success == 1
+        assert duration is not None and duration >= 0
 
     @patch("posthog.tasks.hypercache_verification._run_verification_for_cache")
     def test_skips_when_lock_already_held(self, mock_run_verification: MagicMock) -> None:
